Reworked ground/platform generation
Completely reworked the way that new ground/platform are generated. Now have five arrays for ground prefabs and five arrays for platform prefabs. Each ground/platform array stores a platform based on difficulty.

Platforms are selected from the appropriate array, based on the difficulty of the game. Difficulty is altered by the distance travelled by the player.

Game is working fine, but both ground and platform prefabs will need to be fine tuned for obstacle generation.