Reworked Announcer and Announcer Manager
Reworked the announcers to hold time variables for the time it takes to say "Super Cave Runner" and "D". Announcer Manager altered to take the values from these variables to determine when the sound effects are to be played.

Methods also included to display the title screen, depending on the announcer. Primitive at the moment, but implemented for work to be done in the future.