Working on colour change mechanic
Working on the colour change mechanic. So far, ability implemented into the player character can press a button to change the sprite colour from one of the two determined colours.

Now working on the light gates and light emitters. Sprites made, but will need to code logic to allow them to be generated to the ground and have actions occur when the player comes into contact with an emitter.