Animated character sprites added and implemented
Animated character sprites have now been created and added to the game. Sprites consist of png sprite sheets, which are used to create animations with Unity's Animator system.

Animator file for the player added which controls the animation states of the player sprites. Player can change colour as before, however sprite is changed for the appropriate colour instead of manually applying a colour to the sprite.

Currently, animator for the player controls the state from idle, to running, to jumping and back to running, as well as when the player changes colour. Will need to implement deaths.

Player controller script altered to remove code that changed the sprite colour and new Player Animator Controller script added to link the player sprite with the state of the player.

Fixed the way that difficulty was applied to the game and fixed the result analyser to reflect this change.

All mechanics implemented, these now need fine tuning.