Added camera movement and other improvements
Added camera movement to the project. Player controller now tracks space above the characters head. When point above the players head leaves the top edge of the screen, the camera follows the player character up. When the point above the players head returns back into the main game space, the camera returns to the origin point.

This prevents the character leaving the screen and is always in view for the player. This can also add extra challenge and may make players perform leaps of faith on harder areas of the level.

Background parallax objects have been tweaked slightly. The sprites that cover the areas above the pixel art now has the correct layer orders and are now rendered correctly.

Minor improvements to the background parallax have been implemented.