Best score system added
Best distance system added to the project.

When the game scene starts, the distance text is hidden from view as it is zero.

After the player has played, died, and been given a distance, the best distance is updated, provided the distance is bigger than the distance recorded.

When the game restarts after the best distance has been recorded, the best distance is shown on screen.

Currently toying with the idea of making it behave like a counter that updates in real time, once the a better distance is achieved.