Tweaking light mechanics and player behaviour
Light gates now result in instant death if the player is wearing the wrong colour of armour. Makes the game more challenging and really highlights the change colour mechanic.

Tweaked the best score system to always display the best score, and update in realtime when the player has achieved a better score.

Added dumb hidden secret to UI Distance script.