Main title animated for announcer 01 and 02
The main title is now animated for announcer 1 and 2. "Cave Runner" fades in as before. Once fully in view, the word "Super" slides in as the announcer begins announcing the name of the game. Once announced, the "D" jitters into view as the announcer says "D", then the "X" jitters in as the announcer says "X".

Script added to obtain and display the game version on the title screen.

Fixed bug where sometimes the player would die at a light beam and show the same armour colour as the light beam. This is caused when the player changes armor colour a frame or two before entering the light beam. To fix this, a could of extra paths have been added between the ColourOneDyingLightGate and ColourTwoDyingLightGate states in the animator, where the paths depend on the boolean "Is Colour One". Hopefully this bug won't appear again...

Will need a few more tweaks to the platforms, but programming is more or less complete!