Bug fixing
Attempting to fix a bug where the player can sometimes die in a light beam and the armour is the same colour as the beam. Added a small invincibility time that is activated when the player changes armour, ignoring the light beam for a very small amount of time. Should give the game enough time to change to the appropriate sprite.